+++ libs/ledger-live-common/src/__tests__/test-helpers/environment.ts
@@ -91,6 +91,7 @@ setSupportedCurrencies([
   "telos_evm",
   "coreum",
   "injective",
+  "vechain",
   "casper",
   "neon_evm",
   "lukso",